I was trying to fix the problem of distributing public encryption keys, but have now thought of a better approach, so am reverting. The new approach will be:
The 'Msg' field in 'ClientToClient' will be an interface instead of a byte slice. It will just be a basic shell interface to give me a bit more type safety. Then I can have types that implement it, like 'Encrypted', and 'Key share' etc, and there will be a type switch on the receiving end to work out which one it is.