This commit changes BWT from being purely a web app to an installed desktop app.
The reason for doing this is that I couldn't find a way to do the
web app securely. The security model is that the server cannot be
trusted by the client. The problem with having a web app and encrypting
the data in Javascript is that if the server was malicious it could
steal all the user's data by serving it some malicious Javascript.

It uses Go and Webview to display the GUI. This means that I can use
the excellent Go nacl library for the crypto. The advantage of Webview
is that I can reuse all my Elm front-end code.