Some more progress on using a channel-based approach, but hitting some problems.
The basic issue is that there is no request-response. Messages are
pushed out into the world, and the world pushes messages back in.
So for example, in the goroutine for sending large messages, I tried
to do this:

    ch.cache.to <-CacheGet(draftId)
    rawDraft := <-ch.cache.from.got

So I am asking the cache for some data and waiting for it to come back.
But since the cache could be being asked for other things at the same
time by other goroutines, how do I know that what I get back is
what I want?