Progress on sender, using one-time encryption keys.
It's difficult, this idea of using one-time asymmetric encryption keys. The idea was to improve the forward secrecy. The problem is, how can the recipient authenticate the message, when they don't have the recipient's public encryption key?

I could do encrypt-then-sign, using SealAnonymous, but then there is nothing to prevent an adversary stripping off the signature and adding their own, making it look like the message came from them.