Reverted back to Haskell version.
My main choices of languages that I am familiar with and that have suitable crypto libraries are C, Haskell and Go.

Go and C score well for consistency and simplicity, but badly for immutability. C is bad for concurrency and networking libraries.

Haskell comes out best overall for the things that I care about, but scores low on consistency and simplicity.