Deleted everything but the spec. I've decided to use Go on the front and back end:
I need a cross-platform language. I looked at Haskell, Go, Rust, JS in browser, C and Zig.

+ Haskell is bad for cross-platform. How to run it on Android?

That leaves Go, Rust, JS, C and Zig for the client.

+ I'm slightly doubtful about JS, Rust and to some extent Zig for crypto. They're OK, but not the best and I want the best.

That leaves Go and C.

+ C has no advantages over Go for my situation, and Go has some over C: easier to write, less verbose, easy concurrency and networking, easy cross-platform.

That leaves Go for the client.

Haskell is still good for the server, is a better language, but I think I will use Go there too, just because it's simpler to have the whole thing in one language.

The Go backend will just be a binary that starts up a localhost server for deploying the local web app. So the platform specific wrapper in whatever language will just start the Go binary and tell it's webview widget the URL of the home page. The JS will be the bare bare minimum.